+++ src/components/Stunting/stunting.module.css
@@ -1,0 +1,38 @@
+.wrapper {
+  display: block;
+  padding: 5rem;
+}
+
+.content-main {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+  text-align: justify;
+  text-justify: inter-word;
+}
+
+.img-container {
+  display: flex;
+  gap: 1rem;
+}
+
+p {
+  line-height: 1.5rem;
+}
+
+h3 {
+  padding-bottom: 15px;
+}
+
+h1 {
+  text-align: center;
+  padding-bottom: 20px;
+}
+
+hr {
+  padding-bottom: 20px;
+  border-top: 2.5px solid #00525F;
+  border-bottom: 0px;
+  border-left: 0px;
+  border-right: 0px;
+}
